// Sche_Treatment_Project/src/Models/Model.tsx
export interface Area {
  id: string;
  areaName: string;
  createAt: string;
  updateAt: string;
  supportStatus: string;
  clinics: Clinic[];
}
export interface AreaDTO {
  id: string;
  areaName: string;
  supportStatusId: SupportDTO;
}
export interface PackageEntity {
  id: string;
  packageName: string;
  packagePrice: string;
  createdAt: string;
  updatedAt: string;
  clinicId: Clinic;
  def:boolean;
  supportStatusId: Support;
  packageServices: PackageService[];
}
export interface PackageDTO {
  id: string;
  packageName: string;
  packagePrice: string;
  clinicId: Clinic | undefined;
  def: boolean;
  supportStatusId: string;
  packageServices: PackageServiceDTO[];
}

export interface PackageServiceDTO {
  id: string;
  medicalService: string;
  packageId: string;
  // supportStatusId: string;
}
export interface PackageService {
  id: string;
  medicalService: ServiceEntity;
  packageId: string;
  supportStatusId: Support;
}
export interface ServiceDTO {
  id: string;
  serviceName: string;
  servicePrice: string;
  serviceDescription: string;
  supportStatusId: string;
  clinic: string;
}
export interface ServiceEntity {
  id: string;
  serviceName: string;
  servicePrice: string;
  serviceDescription: string;
  supportStatusId: Support;
  createdAt: string;
  updatedAt: string;
  clinic: Clinic;
}
// ...
export interface DoctorEntity {
  id: string;
  accountPhone: string;
  accountEmail: string;
  accountName: string;
  accountGender: number;
  createAt: string;
  updateAt: string;
  supportRole: Support;
  supportStatus: Support;
  doctorDegree: string;
  doctorRank: string;
  doctorSpecialty: string;
  doctorIntroduce: string;
  doctorExp: string;
  doctorImage: string;
}
// ...
export interface CalendarModel {
  id: string;
  calendarDate: string;
  createAt: string;
  updateAt: string;
  doctor: DoctorEntity;
  groupTime: GroupTime;
}

export interface GroupTime {
  id: string;
  groupTimeDescription: string;
  supports: Support[];
}
export interface Support {
  id: string;
  supportValue: string;
  supportInfo: string;
  createAt: string;
  updateAt: string;
  groupTime: GroupTime;
}
export interface SupportDTO {
  id: string;
  supportValue?: string;
  supportInfo?: string;
}
export interface Clinic {
  id: string;
  clinicName: string;
  createAt: string;
  updateAt: string;
  supportStatus: Support;
  medicalAreaId: Area;
  calendars: CalendarModel[];
}
export interface ClinicDTO {
  id: string;
  clinicName: string;
  supportStatusId: string;
  medicalAreaId: Area | null;
}